remove unsigned

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -7,7 +7,7 @@
  */
 void print_rev(char *s)
 {
-	unsigned long int index;
+	int index;
 	char letter;
 
 	for (index = (strlen(s) - 1); index >= 0; index--)
